Move parser to lib
